snapcraft.yaml: re-organize snapcraft file staging
Previously, all the parts assumed they were run in a subdir
parts/$PART/ of the main git root and used ../../ to access the patches,
etc. but this is not portable to all version of snapcraft and also
doesn't work well with multipass/lxd/launchpad for all the various
places that parts can be located with respect to the main git root.

This patch adds a local-assets part which dumps the sources into
$SNAPCRAFT_STAGE specifically and the other parts pull the files from
there as well.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>